+++ views/public/assets/css/popupImg.css
@@ -1,0 +1,37 @@
+/* Large Image Popup */
+.hideLargeImg{
+    position: absolute;
+    right: 20px;
+    top: 20px;
+    color: white;
+    z-index: 6;
+}
+
+.hideLargeImg:hover {
+    color: red;
+}
+
+.largeImgContainer{
+    position: fixed;
+    display: flex;
+    right: 0;
+    left: 0;
+    top: 5px;
+    margin-right: auto;
+    margin-left: auto;
+    width: 99vw;
+    height: 99vh;
+    background-color: rgba(0, 0, 0, 0.8);
+    z-index: 4;
+}
+
+.viewLargeImage {
+    object-fit: contain;
+    margin: auto;
+    padding: 20px;
+    max-width: 92%;
+    max-height: 90%;
+    z-index: 1;
+    position: relative;
+
+}
