+++ views/public/index.html
@@ -9,6 +9,8 @@
     />
      <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@100;300;400;500;700;900&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="/assets/css/style.css">
+    <!-- Image Fullscreen Popup -->
+    <link rel="stylesheet" href="/assets/css/popupImg.css">
     <title>Faam</title>
   </head>
   <body>
