+++ views/src/components/backgroundCarousel/backgroundCarousel.css
@@ -1,0 +1,16 @@
+
+.backgroundCarousel{
+    z-index: -1;
+    position:fixed;
+}
+
+.profileBackgroundImages{
+    object-fit: cover;
+    width: 100%;
+    max-height: 100%;
+    margin-top: 10px;
+}
+
+
+
+
