+++ views/src/components/post/post.css
@@ -5,13 +5,12 @@
     border: 8px solid rgb(49,49,49);
     background-color: rgba(255,255,255, 0.75);
     transition: 1s;
-
 }
 
-.post:hover{
-    background-color: rgba(255,255,255,1);
-    transition: 1s;
-}
+    .post:hover{
+        background-color: rgba(255,255,255,1);
+        transition: 1s;
+    }
 
 .postWrapper{
     padding: 15px;
@@ -33,23 +32,18 @@
     height: 32px;
     border-radius: 20%;
     object-fit: cover;
-
 }
 
 .postUsername{
     font-style: 15px;
     font-weight: 500;
     margin: 0 10px;
-
 }
 
 
 a:visited, a{
     color: black;
-    
 }
-
-
 
 .postDate{
     font-size: 12px;
@@ -57,7 +51,6 @@ a:visited, a{
 
 .postCenter{
     margin: 20px 0;
-
 }
 
 .postImg{
@@ -66,7 +59,6 @@ a:visited, a{
     max-height: 45vh;
     object-fit: contain;
 }
-
 
 .postBottom{
     display: flex;
@@ -99,9 +91,9 @@ a:visited, a{
     flex-direction: row;
 }
 
-.postCommentText:hover{
-    font-size: 17px;
-}
+    .postCommentText:hover{
+        font-size: 17px;
+    }
 
 .postCommentSpan{
     padding-right: 2px;
@@ -111,22 +103,20 @@ a:visited, a{
     font-size: 14px;
 }
 
-
-.postDropDown:hover{
-    cursor: pointer;
-    color:deepskyblue;
-}
-
-
-.postDelete:hover{
-    cursor: pointer;
-    color: red;
-}
-
-
+/* Post Dropdown */
 .postTopRight {
     display: flex;
 }
+
+    .postDropDown:hover{
+        cursor: pointer;
+        color:deepskyblue;
+    }
+
+    .postDelete:hover{
+        cursor: pointer;
+        color: red;
+    }
 
 .editPostButton{
     border:none;
@@ -140,17 +130,17 @@ a:visited, a{
     font-size: 15px;
 }
 
-.editPostButton:hover {
-    background-color: rgb(25,230,148);
-}
+    .editPostButton:hover {
+        background-color: rgb(25,230,148);
+    }
 
-.cancelEditPostButton{
-    background-color: rgb(255, 0, 102);
-}
+    .cancelEditPostButton{
+        background-color: rgb(255, 0, 102);
+    }
 
-.cancelEditPostButton:hover {
-    background-color: rgb(230, 25, 107);
-}
+    .cancelEditPostButton:hover {
+        background-color: rgb(230, 25, 107);
+    }
 
 
 .editPostTextInput{
