--- views/src/components/profileTop/profileTop.css
@@ -1,82 +1,0 @@
-
-
-/* 
-.saveCoverImgButton {
-    width: 32px;
-    height: 32px;
-    border-radius: 50%;
-    position: absolute;
-    background-color: #77BC3F;
-    border: none;
-    bottom: 55px ;
-    left: 60px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 24px;
-    z-index: 1;
-}
-
-.saveCoverImgButton:hover{
-    background-color: #75cc2f;
-}
-
-.coverPicCancelImg{
-    min-width: 32px;
-    min-height: 32px;
-    background-color: #ff0000; 
-    border-radius: 50%;
-    position: absolute;
-    bottom: 55px ;
-    left: 15px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 24px;
-    z-index: 1;
-}
-
-.coverPicCancelImg:hover{
-    background-color: #ff4040;
-} */
-
-/* .profileCoverCameraIcon {
-    width: 30px;
-    height: 30px;
-    background-color: lightgray;
-    border-radius: 50%;
-    bottom: 58px ;
-    left: 15px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 12px;
-
-    display: none;
-}
-
-.profileCoverCameraIcon:hover {
-    background-color: lightgreen;
-} */
-
-
-/* Carousel */
-
-.profileCover{
-    z-index: -1;
-    position:fixed;
-}
-
-.profileBackgroundImages{
-    object-fit: cover;
-    width: 100%;
-    /* Max Height affects height and works */
-    max-height: 100%;
-    margin-top: 10px;
-    /* margin-right: 50%; */
-    /* margin-left: 20px; */
-}
-
-
-
-
