+++ views/src/components/rightbar/rightbar.css
@@ -3,19 +3,16 @@
     background-color: transparent;
     height: calc(100vh);
     top: 0;
-    position: sticky !important;
+    position: sticky;
     background-color: rgba(255,255,255, 0.85);
     transition: 0.25s;
-
 }
-
-.rightbar:hover{
-    background-color: rgba(255,255,255, 1);
-    
-}
+    .rightbar:hover{
+        background-color: rgba(255,255,255, 1);   
+    }
 
 .profileRightbarMargin{
-    position:fixed !important;
+    position:fixed;
     right: 0;
     left:auto;
     width: 425px;
@@ -23,13 +20,12 @@
 
 .rightbarWrapper{
     padding: 20px 20px 0 20px;
-
 }
 
 .rightbarTitle{
     margin-bottom: 20px;
 }
-
+ 
 .rightbarFriendList{
     padding: 0;
     margin: 0;
@@ -50,9 +46,7 @@
 
 .rightbarInfo {
     margin-bottom: 30px;
-
 }
-
 
 .rightbarInfoItem {
     margin-bottom: 10px;
@@ -88,8 +82,6 @@
     object-fit: cover;
     border-radius: 5px;
 }
-
-
 
 .rightbarFollowButton{
     margin-top: 30px;
@@ -133,22 +125,29 @@ display: flex;
     background-color: rgba(236, 236, 236, 1);
     background-color: white;
     border-radius: 8px;
-    max-height: 380px;
+    max-height: 180px;
     overflow: scroll;
-
     -webkit-box-shadow: 0px 1px 7px -1px rgba(0,0,0,0.37); 
     box-shadow: 0px 1px 7px -1px rgba(0,0,0,0.37);
-
     display: grid;
     grid-template-columns: repeat(2, 1fr);
     gap: 5px;
     row-gap: 10px;
     padding:10px 5px 10px 5px;
     overflow-x: hidden;
-    
 }
 
+.photoAlbumHeaderContainer{
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    margin-right: 15px;
+    margin-left: 5px;
+}
 
+.rightbarTitleSeeAll:hover{
+    color: rgb(0,255,153);
+}
 
 .sidebarAllUserPhotos{
     object-fit: cover;
@@ -173,54 +172,14 @@ display: flex;
 }
 
 
-/* Large Image Popup */
-.hideLargeImg{
-    position: absolute;
-    right: 20px;
-    top: 20px;
-    color: white;
-    z-index: 6;
-}
-
-.hideLargeImg:hover {
-    color: red;
-}
-
-.largeImgContainer{
-    position: fixed;
-    display: flex;
-    right: 0;
-    left: 0;
-    top: 5px;
-    margin-right: auto;
-    margin-left: auto;
-    width: 99vw;
-    height: 99vh;
-    background-color: rgba(0, 0, 0, 0.8);
-    z-index: 4;
-}
-
-.viewLargeImage {
-    object-fit: contain;
-    margin: auto;
-    padding: 20px;
-    max-width: 92%;
-    max-height: 90%;
-    z-index: 1;
-    position: relative;
-
-}
-
 .allUsersHeader {
     margin-top: 15px;
     margin-bottom: 15px;
-
 }
 
 .rightbarHr{
     margin-top: 10px;
     margin-bottom: 10px;
- 
 }
 
 .rightbarProfileHr{
@@ -231,7 +190,6 @@ display: flex;
     padding: 5px 5px;
     max-height: 75vh;
     object-fit:scale-down;
-
 }
 
 .rightbarCarouselContainer{
@@ -290,17 +248,16 @@ display: flex;
 
 
 
-/* Profile Right Top Image */
+/* Profile Image */
 
 .profileInfo{
     right: 0;
     display: flex;
-    justify-content: flex-end;
-    
+    justify-content: flex-end;  
 }
 
 
-.profileInfoName{
+.rightbarprofileInfoName{
     box-sizing: border-box;
     background-color: white;
     border: 8px solid black;
@@ -309,7 +266,7 @@ display: flex;
     z-index: 1;
     position:relative;
     top: -75px;
-    font-size: 48px;
+    font-size: 24px;
     width: fit-content;
     margin-right: auto;
     margin-left: auto;
@@ -338,11 +295,10 @@ display: flex;
     z-index: 1;
     transition: 0.25s;
 }
-
-.profileUserImg:hover{
-    padding: 5px;
-    background-color: lightgreen;
-}
+    .profileUserImg:hover{
+        padding: 5px;
+        background-color: lightgreen;
+    }
 
 .profilePicCancelImg {
     min-width: 32px;
@@ -356,10 +312,9 @@ display: flex;
     font-size: 24px;
     z-index: 12;
 }
-
-.profilePicCancelImg:hover{
-    font-weight: bold;
-}
+    .profilePicCancelImg:hover{
+        font-weight: bold;
+    }
 
 
 .saveProfileImgButton {
@@ -367,10 +322,9 @@ display: flex;
     border: none;
     font-size: 24px;
 }
-
-.saveProfileImgButton:hover{
-    font-weight: bold;
-}
+    .saveProfileImgButton:hover{
+        font-weight: bold;
+    }
 
 .saveNewImgContainer{
     width: fit-content;
@@ -378,8 +332,7 @@ display: flex;
     display: flex;
     justify-content: center;
     margin-left: auto;
-    margin-right: auto;
-    
+    margin-right: auto;   
 }
