+++ views/src/components/settingsCenter/settingsCenter.css
@@ -1,11 +1,9 @@
 .settingsCenter{
-    /* width: 33%; */
     flex: 4;
     margin-top: 10px;
     margin-left: 20px;
     margin-right: 20px;
     display: flex;
-    
 }
 
 .middleSettings{
@@ -66,4 +64,20 @@
 .settingsGoBack{
     text-decoration: none;
     padding-bottom: 5px;
+}
+
+.settingsProfileInfoName{
+    box-sizing: border-box;
+    background-color: white;
+    border: 8px solid black;
+    padding: 5px;
+    padding-top: 10px;
+    z-index: 1;
+    position:relative;
+    top: -75px;
+    font-size: 48px;
+    width: fit-content;
+    margin-right: auto;
+    margin-left: auto;
+
 }
