+++ views/src/components/share/share.css
@@ -39,13 +39,9 @@
     width: 80%;
     background-color: rgb(245,245,245);
 }
-
-.shareInput:focus{
-    outline:none;
-    /* height: 40px;
-    font-size: larger; */
-     
-}
+    .shareInput:focus{
+        outline:none;
+    }
 
 .shareHr{
     margin: 20px;
@@ -92,24 +88,21 @@
     color: Black;
     font-size: 16px;
 }
-
-.shareButton:hover {
-    background-color: rgb(25,230,148);
-}
+    .shareButton:hover {
+        background-color: rgb(25,230,148);
+    }
 
 .shareImgContainer{
     padding: 0 20px 10px 20px;
     position:relative;
     display: flex;
-    justify-content: center;
-    
+    justify-content: center;    
 }
 
 .shareImg{
     max-width:100%;
     object-fit: contain;
     max-height: 75vh;
-
 }
 
 .shareCancelImg{
@@ -122,7 +115,6 @@
     margin-top: 10px;
     margin-right: 10px;
 }
-
-.shareCancelImg:hover {
-    color: red;
-}
+    .shareCancelImg:hover {
+        color: red;
+    }
