+++ views/src/components/topbar/topbar.css
@@ -10,29 +10,7 @@
     top: 0;
     right: 0;
     justify-content: space-around;
-
 }
-
-
-/* the flex:# divides the topbar into units. Right now it's divided into 12 units */
-
-
-
-.logo{
-    font-size: 24px;
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    
-}
-
-.logoImg {
-    margin-top: 10px;
-    width: 90%;
-}
-
-
 
 .topbarRight{
     display: flex;
@@ -50,7 +28,6 @@
     justify-content: space-evenly;
     align-items:center;
     margin-right: 10px;
-
 }
 
 .topbarLink{
@@ -59,13 +36,11 @@
     color: black;
     box-sizing: border-box;
     padding: 4px;
-    
 }
 
 .topbarLink:hover{
     font-weight: bold;
 }
-
 
 .topbarIcons{
     display:flex;
@@ -77,7 +52,6 @@
     cursor: pointer;
     position: relative;
 }
-
 
 .topbarIconBadge {
     width: 15px;
@@ -92,19 +66,15 @@
     align-items: center;
     justify-content: center;
     font-size: 12px;
-
 }
-
 
 .topbarImg{
     width: 38px;
     height: 38px;
     border-radius: 20%;
-    /* Picture looks better with object-fit cover */
     object-fit: cover;
     cursor: pointer;
 }
-
 
 #logoutButton {
     width: 100%;
@@ -121,7 +91,6 @@
 #logoutButton:hover{
     font-weight: bold;
 }
-
 
 .notificationContainer{
     position: absolute;
