+++ views/src/pages/about/about.css
@@ -7,7 +7,7 @@
     justify-content: center;
 }
 
-.logoWrapper{
+.aboutlogoWrapper{
     margin-left: auto;
     margin-right: auto;
     background-color: white;
@@ -44,9 +44,6 @@
     border-radius: 20px;
     min-width: 300px;
 }
-
-
-
 
 .aboutH2{
     font-size: 48px;
@@ -125,48 +122,31 @@
 /* Media */
 
 @media screen and (max-width: 1425px) {
-    
-
 
      .aboutSidebarLeft{
         display: none;
     }
 
-
     .aboutTextWrapper{
         width: 70%;
     }
-
 }
 
 
 @media screen and (max-width: 1000px) {
     
-
-
     .aboutSidebar{
        display: none;
    }
-
-
    .aboutTextWrapper{
        width: 90%;
    }
-
 }
-
 
 
 @media screen and (max-width: 750px) {
 
     .aboutLogo{
         width: 100%;
-    }
-
-    
-}
-
-
-@media screen and (max-width: 425px) {
-    
+    }    
 }
