+++ views/src/pages/photoAlbum/photoAlbum.css
@@ -1,8 +1,7 @@
 .photoAlbumContainer {
     display: flex;
     background-color: rgb(236,236,236);
-    max-height: 100vh;
-    
+    max-height: 100vh; 
 }
 .photoAlbum {
     flex: 5.5;
@@ -18,19 +17,5 @@
     background-color:#fff;
     box-sizing: border-box;
     border: 10px solid rgb(49,49,49);
-
-
 }
 
-.photoAlbumHr{
-    width: 98%;
-    margin-left: auto;
-    margin-right: auto;
-    margin-bottom: 15px;
-}
-
-
-.rightbarPhotoAlbum{
-    position: sticky;
-    overflow-y: scroll;
-}
