+++ views/src/pages/profile/Profile.jsx
@@ -2,7 +2,7 @@ import "./profile.css"
 import Sidebar from "../../components/sidebar/Sidebar"
 import Rightbar from "../../components/rightbar/Rightbar"
 import Feed from "../../components/feed/Feed"
-import ProfileTop from "../../components/profileTop/ProfileTop"
+import BackgroundCarousel from "../../components/backgroundCarousel/BackgroundCarousel"
 import { useEffect, useState } from 'react'
 import axios from "axios"
 import { useParams } from "react-router"
@@ -28,10 +28,10 @@ export default function Profile() {
     return (
         <>
         <div className="profile">
-            <Sidebar className="profileSidebar" />
+            <Sidebar className="profilePageSidebar" />
             <div className="profileRight">
                 <div className="profileRightTop">
-                    <ProfileTop />                
+                    <BackgroundCarousel />                
                 </div>
                 <div className="profileRightBottom">
                     <Feed username={username} />
