+++ views/src/pages/settings/settings.css
@@ -1,17 +1,5 @@
 .settingsWrapper {
     display: flex;
-    /* align-items: center; */
-    justify-content: center;
-    
-}
-
-.leftSettingsSidebar{
-    /* width: 8vw; */
-    
-}
-
-.rightSettingsSidebar {
-    /* width: 85vw; */
-    
+    justify-content: center;    
 }
